remove features enabled in the newer Perl version
Given code like:

    use v5.20;
    use feature qw( lexical_subs signatures );
    no warnings 'experimental::lexical_subs', 'experimental::signatures';

When bumping to v5.26, lexical_subs is enabled by default,
but signatures are still experimental, so it should become:

    use v5.26;
    use feature 'signatures';
    no warnings 'experimental::signatures';

In 5.36, signatures are not experimental any more, so it should become:

    use v5.36;

This patch also handles features enabled with the `experimental`
module, which also disables the experimental warnings.

The private `_ppi_list_to_perl_list` function is turning a list
of PPI tokens back into a Perl data structure. This obviously only
work for static expressions, but that's the case for what we're
dealing with in our case.