temporarily accommodate for Perl Critic rules
Perl::Critic::Policy::Modules::RequireExplicitPackage wants the package
line to be the first significant line in a module.

It's a trivial patch to accommodate for "line 1 semantics", which
has been sent upstream. Until that is merged, move the `use VERSION`
line to the top, or after the package line if it's already the first
significant line.

The Perl-Critic patch is: https://github.com/Perl-Critic/Perl-Critic/pull/1070