handle specific features before removing obsoleted ones
In this case, the issue was triggered by `bareword::filehandles`: if
the old code had `no bareword::filehandles`, that would be turned to
`no feature 'bareword_filehandles'`. However since the cleanup of `no
feature` for disabled obsolete features was done before the upgrade from
`no bareword::filehandles` to the `no feature 'bareword_filehandles`,
it was too late to remove the new added `no` statement.

This patch handles the specific features before removing the disabled
ones.