compute everything we know about features in a simple hash
There's some data that can't be obtained from the data structures in
the `feature` module, for example when a feature first appeared.
Other features are not even listed in those data structures.

This data structure should help make the code more generic.

This was written on a Perl 5.36, so only document features up to that
version.