read the feature data from a table in the __DATA__ section
This means that the module's knowledge of Perl features does not depend
on the version of the perl binary that is running it. It follows that
the result of bumping does not depend of the version of perl binary
running the module. And so, the tests results should be consistent on
every perl version.

This commit also introduces a script that generates the tabular data for
all the features. Whenever a new Perl will be released, just update the
__DATA__ section of the module.

The version of the feature set also serves as the maximum value for the
`version` attribute. The module will refuse to bump after that number.