do not drop features that the perl running the code does not know about
We have a database of information that knows about all features, so
the linter should be able to run on an older perl (say perl5.16.3)
and be asked to bump some code to a later version of Perl (say v5.20).

This makes the "known to perl" logic irrelevant.

This code:

    use feature 'signatures';

will die with any perl before perl5.20.0, because the binary does not
know about the feature itself.  However, that should not prevent a
perl5.20.0 to bump it to any version of the language.

The above code will work on any perl that knows about the feature, and
only enable signatures, but not `say`, or any other post-v5.10 feature.

When writing code like this:

    use v5.10;
    use feature 'signatures';

one is opting into a version of the Perl language that is not described
by any specific feature bundle. However, any perl that knows about the
`signatures` feature will know how to run it.