create bump_ppi_safely and build bump_safely and bump_file_safely on top
The interface is a little weird: it returns a single value in scalar
context, and two values in list context.

The second value is needed so that it can be the actual return
value for bump_file_safely. I don't expect users of bump_ppi
or bump_safely to really need it.