get rid of wantarray in bump_ppi_safely
The private _bump_ppi_safely returns an arrayref with both
values (modified PPI document and target version), and the
public methods use that and return the value they have to.

This slighly simplifies the code, but more importantly,
avoids bubbling the weird scalar/list context differences
up to the module's public interface.